first commit for refacto visibility
